requirements.txt
youtube-transcript-api==0.6.1
yt-dlp==2023.12.30
sentence-transformers==2.2.2
chromadb==0.4.21
python-dotenv==1.0.0
openai==1.6.1
transformers==4.36.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
click==8.1.7
tqdm==4.66.1
langchain==0.1.0
langchain-community==0.0.10
gpt4all==2.8.2
huggingface_hub==0.15.1
youtube-transcript-api==1.1.1
yt_dlp==2025.6.30
chromadb==1.0.15